backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.4.0
beautifulsoup4==4.12.2
bitarray==3.7.2
certifi==2025.10.5
charset-normalizer==3.4.4
ckzg==1.0.2
cytoolz==1.1.0
distro==1.9.0
dnspython==2.4.2
eth-abi==4.2.1
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.1
exceptiongroup==1.3.0
feedparser==6.0.10
frozenlist==1.8.0
groq==0.33.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lru-dict==1.2.0
motor==3.3.2
multidict==6.7.0
oauthlib==3.3.1
parsimonious==0.9.0
propcache==0.4.1
protobuf==6.33.0
py-ecc==8.0.0
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pymongo==4.6.0
python-dotenv==1.0.0
python-telegram-bot==22.5
pyunormalize==17.0.0
redis==5.0.1
referencing==0.36.2
regex==2025.10.22
requests==2.31.0
requests-oauthlib==1.3.1
rlp==4.1.0
rpds-py==0.27.1
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.8
toolz==1.1.0
tweepy==4.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
web3==6.11.3
websockets==15.0.1
yarl==1.22.0
